requirements.txt
websocket-client[optional]
pyyaml
telethon
python_socks
async_timeout
coloredlogs
websockets
aiohttp[speedups]==3.8.6
ffmpy3